+++ examples/get_moderation.rs
@@ -1,3 +1,4 @@
+use futures::TryStreamExt;
 use twitch_api2::{
     helix::moderation::{GetBannedEventsRequest, GetBannedUsersRequest, GetModeratorEventsRequest},
     HelixClient,
@@ -41,7 +42,8 @@ async fn run() -> Result<(), Box<dyn std::error::Error + Send + Sync + 'static>>
         "{:?}",
         client
             .get_moderators_in_channel_from_id(broadcaster_id, &token)
-            .await?
+            .try_collect::<Vec<_>>()
+            .await?,
     );
 
     println!("====Last 20 Moderator Events====");